refactor(hmpb): rename & relocate timing mark paper fixtures
Instead of putting the fixture generator into the `ballot_templates`
directory with the actual ballot templates, move them into their own
`timing_mark_paper` folder a la `all_bubble_ballot`. Also changes the
name from "timing mark grid-only" to "timing mark paper".
